Integration test failure fix.
